// src/app/_models/enums.ts
export enum Enums {
    ADMIN = 'ADMIN',
    SKITEACHER = 'SKITEACHER',
    RACE = 'RACE',
    INSTRUCTOR = 'INSTRUCTOR',
    SKITEAM = 'SKITEAM',
    CONTACTPERSON = 'CONTACTPERSON',
    DEFAULT = 'DEFAULT'
}

export enum Proficiency {
    LILA = 'LILA',
    ORANGE = 'ORANGE',
    GRUEN = 'GRÜN',
    GELB = 'GELB',
    DEFAULT = 'DEFAULT'

}